Add files via upload
The dataset (Amazon dataset)  for this project was gotten from Kaggle an open source platform where you can get dataset for analysis. After cleaning the dataset I was able to draw some meaningful insight from it and I was able to visualize it and create a dashboard. Now, from the dashboard there are insights  which would help the company to take proactive business decisions that can improve the growth of the business and also how they can satisfy their customers.