Add files via upload
In this Project I help a Sales store to analyze their data from the sales they made in 2018 so as to help them improve in their services. I was also able to create a dashboard which will help the stakeholders to get visual information about the progress and setback in their business and will also help them make decision for improvement.